Improve useDummyEnvFiles.sh messaging (#6)
* Fix messaging on useDummyEnvFiles.sh to be a bit more clear and less foreboding.
* Add a force flag to allow it to pass on circle ci and a help command to learn how to use the script.

+++ scripts/useDummyEnvFiles.sh
@@ -1,8 +1,39 @@
 #!/bin/bash
 
-echo "This script must be run from the top-level guardian-bio-auth directory. If you run it from anywhere else, it will fail."
+# Parse command line parameters, if any
+HELP=false
+FORCE=false
+while [[ $# -gt 0 ]]
+do
+  if [[ "$1" == "-h" || "$1" == "--help" ]]
+  then
+    HELP=true
+  elif [[ "$1" == "-f" || "$1" == "--force" ]]
+  then
+    FORCE=true
+  fi
+  shift
+done
 
-cp bioanalyzer_service/dummy.env bioanalyzer_service/.env
-cp identity_service/dummy.env identity_service/.env
-cp identity_intelligence_db/dummy.env identity_intelligence_db/.env
-cp identity_template_db/dummy.env identity_template_db/.env
+# Execute. If help is requested, don't actually execute the script.
+if [[ "$HELP" == true ]]
+then
+  echo "Apply dummy environment variables to each of the repos contained in guardian-bio-auth"
+  echo ""
+  echo "USAGE:  useDummyEnvFiles.sh [OPTIONS]"
+  echo ""
+  echo "Options:"
+  echo "-h, --help     Display this help message"
+  echo "-f, --force    Force execute the useDummyEnvFiles.sh script without checking the current working directory"
+elif [[ "$FORCE" == true || "$PWD" == */guardian-bio-auth ]]
+then
+  cp bioanalyzer_service/dummy.env bioanalyzer_service/.env
+  cp identity_service/dummy.env identity_service/.env
+  cp identity_intelligence_db/dummy.env identity_intelligence_db/.env
+  cp identity_template_db/dummy.env identity_template_db/.env
+  echo "Success"
+  exit 0
+else
+  echo "This script must be run from the top-level guardian-bio-auth directory. You are currently running it from: $PWD"
+  exit 1
+fi
